TRI-48 해당 API 접근 허용. 시큐리티 설정
